% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Int vector of length 1 (Alabama = 1, Wyoming = 50, etc)}

\item{year}{Int vector of length 1 of the year desired (2013, 2014, etc)}
}
\value{
A png file with the number of observations plotted on a state map
}
\description{
Function to read data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System.
and plot the of observations for the given year on a png state map
based on latitude and longitude of the observations.
}
\note{
parameters 'state.num' and 'year' must each be a length 1 vector
}
\examples{
 \dontrun{
  fars_map_state(1,2013)
  fars_map_state(12,2014)
 }

}
