% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{filename of data file}
}
\value{
A tibble: <number of rows> x 50 cols
}
\description{
Function to read a semicolon (;) delimited file (including csv, tsv, .gz, .bz2, or .zip)
into a tibble from data from the US National Highway Traffic Safety Administration's Fatality Analysis
Reporting System. SOURCE: http://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)
}
\note{
Using a non-existant filename will result in a trapped error.
}
\examples{
 \dontrun{
  fars_read("C://User//data.zip")
 }

}
