% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{Integer list or vector of the year(s) desired}
}
\value{
A tibble: <number of rows> x 2 cols (MONTH and year)
}
\description{
Function to read data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System.
by individual file year, placing into a tibble (dataframe)
}
\note{
Requires package 'magrittr'.
}
\examples{
 \dontrun{
  fars_read_years(2013)
  fars_read_years((2013:2015))
  fars_read_years(c(2013:2015))
 }

}
