% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{year}{Integer list or vector of the year(s) desired}
}
\value{
A tibble: 12 x <number of years> (MONTH, year1, year2, ...)
}
\description{
Function to read data from the US National Highway Traffic Safety
Administration's Fatality Analysis Reporting System.
and summarize number of observations for each month over one or more years
into a tibble (dataframe) 12 x <number of years>
}
\note{
Requires package 'magrittr'.
}
\examples{
 \dontrun{
  fars_summarize_years(2013)
  fars_summarize_years((2013:2015))
  fars_summarize_years(c(2013:2015))
 }

}
